requirements.txt
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.3
idna==3.11
jaraco.classes==3.4.0
jaraco.context==6.1.0
jaraco.functools==4.4.0
jeepney==0.9.0
keyring==25.7.0
more-itertools==10.8.0
Nuitka==2.7.11
ordered-set==4.1.0
patchelf==0.17.2.4
pycparser==2.23
PyQt5-Qt5==5.15.18
PyQt5_sip==12.17.2
PySide6==6.10.1
PySide6_Addons==6.10.1
PySide6_Essentials==6.10.1
requests==2.32.5
SecretStorage==3.5.0
shiboken6==6.10.1
sqlcipher3-binary==0.6.0
urllib3==2.6.2
zstandard==0.25.0
